Moved rest of the code related to analytics from Collect to FirebaseAnalytics
