+++ collect_app/src/main/java/org/odk/collect/android/analytics/Analytics.java
@@ -5,4 +5,6 @@ public interface Analytics {
     void logEvent(String category, String action);
 
     void logEvent(String category, String action, String label);
+
+    void setAnalyticsCollectionEnabled(boolean isAnalyticsEnabled);
 }
