+++ collect_app/src/main/java/org/odk/collect/android/preferences/IdentityPreferences.java
@@ -21,11 +21,17 @@ import android.preference.CheckBoxPreference;
 import org.odk.collect.android.R;
 import org.odk.collect.android.application.Collect;
 import org.odk.collect.android.utilities.MultiClickGuard;
+import org.odk.collect.android.analytics.Analytics;
+
+import javax.inject.Inject;
 
 import static org.odk.collect.android.preferences.GeneralKeys.KEY_ANALYTICS;
 import static org.odk.collect.android.preferences.PreferencesActivity.INTENT_KEY_ADMIN_MODE;
 
 public class IdentityPreferences extends BasePreferenceFragment {
+
+    @Inject
+    Analytics analytics;
 
     public static IdentityPreferences newInstance(boolean adminMode) {
         Bundle bundle = new Bundle();
@@ -41,6 +47,7 @@ public class IdentityPreferences extends BasePreferenceFragment {
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         addPreferencesFromResource(R.xml.identity_preferences);
+        Collect.getInstance().getComponent().inject(this);
 
         findPreference("form_metadata").setOnPreferenceClickListener(preference -> {
             if (MultiClickGuard.allowClick(getClass().getName())) {
@@ -58,7 +65,7 @@ public class IdentityPreferences extends BasePreferenceFragment {
 
         if (analyticsPreference != null) {
             analyticsPreference.setOnPreferenceClickListener(preference -> {
-                Collect.getInstance().setAnalyticsCollectionEnabled(analyticsPreference.isChecked());
+                analytics.setAnalyticsCollectionEnabled(analyticsPreference.isChecked());
                 return true;
             });
         }
