+++ collect_app/src/test/java/org/odk/collect/android/widgets/QuestionWidgetTest.java
@@ -24,6 +24,7 @@ import org.odk.collect.android.audio.Clip;
 import org.odk.collect.android.formentry.media.AudioHelperFactory;
 import org.odk.collect.android.formentry.questions.QuestionDetails;
 import org.odk.collect.android.injection.config.AppDependencyModule;
+import org.odk.collect.android.preferences.GeneralSharedPreferences;
 import org.odk.collect.android.support.MockFormEntryPromptBuilder;
 import org.odk.collect.android.support.RobolectricHelpers;
 import org.odk.collect.android.support.TestScreenContextActivity;
@@ -95,7 +96,7 @@ public class QuestionWidgetTest {
             }
 
             @Override
-            public Analytics providesAnalytics(Application application) {
+            public Analytics providesAnalytics(Application application, GeneralSharedPreferences generalSharedPreferences) {
                 return analytics;
             }
         });
